backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.6.0
redis>=5.0.0
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
pymongo>=4.8.0